Move translator file
